Projectile Damage
Added:
- Projectile Damage
- Made a lot of not
- Made some file changes
