+++ Seize Command/Assets/Scripts/Abstract.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 8ab59ee05fbad6b4db1773538f4c77de
+folderAsset: yes
+timeCreated: 1510340869
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
