+++ Seize Command/Assets/Scripts/Abstract/AbstractAttackManager.cs.meta
