+++ Seize Command/Assets/Scripts/Interfaces.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: e0137374f1413814bb382847ca70189a
+folderAsset: yes
+timeCreated: 1510340884
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
