+++ Seize Command/Assets/Scripts/Interfaces/IFireable.cs
