--- Seize Command/Assets/Scripts/PhysicalProjectile.cs
@@ -1,16 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-[RequireComponent(typeof(Rigidbody2D))]
-public class PhysicalProjectile: MonoBehaviour
-{
-    [SerializeField] float projectileSpeed;
-    Rigidbody2D rb;
-
-    void Start()
-    {
-        rb = GetComponent<Rigidbody2D>();
-        rb.velocity = transform.up * projectileSpeed;
-    }
-}
