+++ Seize Command/Library/shadercompiler-UnityShaderCompiler.exe0.log
@@ -1,2 +1,4 @@
 Base path: C:/Program Files/Unity/Editor/Data
 Cmd: initializeCompiler
+
+Quitting shader compiler process
